crm_accounting/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyodbc==4.0.39
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
aiofiles==23.2.1
pydantic==1.10.12
bcrypt==4.0.1  # Добавлена явная версия bcrypt